refactor: 페이지들을 pages 디렉터리로 모음
