+++ front-end/legeno-around-here/src/components/pages/PostDetail.js
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import Typography from '@material-ui/core/Typography';
 
-import OutBox from './OutBox';
-import { getAccessTokenFromCookie } from '../util/TokenUtils';
-import { findPost } from './postdetail/PostFinder';
-import Bottom from './Bottom';
-import TopBar from './postdetail/TopBar';
-import Image from './postdetail/Image';
+import OutBox from '../OutBox';
+import { getAccessTokenFromCookie } from '../../util/TokenUtils';
+import { findPost } from '../postdetail/PostFinder';
+import Bottom from '../Bottom';
+import TopBar from '../postdetail/TopBar';
+import Image from '../postdetail/Image';
 
 const Line = styled.div`
   width: 100%;
